requirements.txt
numpy==1.19.1
streamlit==0.64.0
pandas==1.1.0
python_dateutil==2.8.1
